--- database/src/main/java/by/itacademy/database/entity/BaseEntity.java
@@ -1,7 +1,0 @@
-package by.itacademy.database.entity;
-
-import java.io.Serializable;
-
-public interface BaseEntity<T extends Serializable> {
-    T getId();
-}
