--- database/src/main/java/by/itacademy/database/entity/Manufacturer.java
@@ -1,42 +1,0 @@
-package by.itacademy.database.entity;
-
-import java.util.HashSet;
-import java.util.Set;
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Data
-@EqualsAndHashCode(exclude = "part")
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString(exclude = "part")
-@Entity
-@Table(name = "manufacturer", schema = "doc_service")
-public class Manufacturer implements BaseEntity<Integer> {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column(name = "name", unique = true, nullable = false)
-    private String name;
-
-    @OneToMany(mappedBy = "manufacturer", cascade = CascadeType.ALL, orphanRemoval = false)
-    private Set<Part> part = new HashSet<>();
-
-    public Manufacturer(String name) {
-        this.name = name;
-    }
-}
-
