--- database/src/main/java/by/itacademy/database/entity/Role.java
@@ -1,40 +1,0 @@
-package by.itacademy.database.entity;
-
-import java.util.HashSet;
-import java.util.Set;
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-
-@Data
-@ToString(exclude = {"id", "user"})
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "role", schema = "doc_service")
-public class Role implements BaseEntity<Integer> {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column(name = "role", unique = true, nullable = false)
-    private String role;
-
-    @OneToMany(mappedBy = "role", cascade = CascadeType.ALL, orphanRemoval = true)
-    private Set<User> user = new HashSet<>();
-
-    public Role(String role) {
-        this.role = role;
-    }
-}
