+++ database/src/main/java/com/itacademy/database/entity/BaseEntity.java
@@ -1,0 +1,7 @@
+package com.itacademy.database.entity;
+
+import java.io.Serializable;
+
+public interface BaseEntity<T extends Serializable> {
+    T getId();
+}
