+++ database/src/main/java/com/itacademy/database/entity/Department.java
@@ -1,0 +1,39 @@
+package com.itacademy.database.entity;
+
+import java.util.HashSet;
+import java.util.Set;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Data
+@ToString(of = "name")
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "department", schema = "doc_service")
+public class Department implements BaseEntity<Integer> {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Column(name = "name", unique = true, nullable = false)
+    private String name;
+
+    @OneToMany(mappedBy = "departments", cascade = CascadeType.ALL, orphanRemoval = true)
+    private Set<UserDetail> userDet = new HashSet<>();
+
+    public Department(String name) {
+        this.name = name;
+    }
+}
